Updated to reflect actual push back behavior instead of dropping transfers
Originally commented that the Timer class based methods dropped some of the transfers which is incorrect.  After observing the data, all transfers were received by the Teensy (20 pulses appear) but not all of the transfers execute before the next transfer should have occurred (transfers are pushed back, not dropped entirely).